+++ RELEASES.md
@@ -7,6 +7,7 @@
 ### Usability
 
 - better decoupled `InputMap` and `ActionState`, providing an `InputMap::which_pressed` API and allowing `ActionState::update` to operate based on any `HashSet<A: Actionlike>` of pressed virtual buttons that you pass in
+- `InputMap` now uses a collected `InputStreams` struct in all of its methods
 
 ### Bug fixes